update to the latest sveltekit
