--- example/svelte-kit/package.json
@@ -1,15 +1,0 @@
-{
-	"name": "svelte-kit",
-	"type": "module",
-	"version": "0.0.1",
-	"scripts": {
-		"dev": "svelte-kit dev",
-		"build": "svelte-kit build",
-		"preview": "svelte-kit preview"
-	},
-	"devDependencies": {
-		"@sveltejs/kit": "next",
-		"svelte": "^3.49.0",
-		"svelte-preprocess-cssmodules": "workspace:*"
-	}
-}
