+++ example/sveltekit/src/routes/+page.svelte
