+++ example/sveltekit/static/favicon.png
