+++ example/sveltekit/svelte.config.js
@@ -1,0 +1,24 @@
+import adapter from '@sveltejs/adapter-auto';
+import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
+import { cssModules } from 'svelte-preprocess-cssmodules';
+
+/** @type {import('@sveltejs/kit').Config} */
+const config = {
+	// Consult https://kit.svelte.dev/docs/integrations#preprocessors
+	// for more information about preprocessors
+	preprocess: [
+		cssModules({
+			includePaths: ['./'],
+		}),
+		vitePreprocess(),
+	],
+
+		kit: {
+			// adapter-auto only supports some environments, see https://kit.svelte.dev/docs/adapter-auto for a list.
+			// If your environment is not supported, or you settled on a specific environment, switch out the adapter.
+			// See https://kit.svelte.dev/docs/adapters for more information about adapters.
+			adapter: adapter()
+		}
+};
+
+export default config;
